requirements.txt
backoff==2.2.1
InquirerPy==0.3.4
lxml==4.9.3
openai
jsonlines
numpy
protobuf
BeautifulSoup4
toml
aioconsole
python-dotenv
gymnasium
Pillow
evaluate
types-tqdm
tiktoken
aiolimiter
beartype==0.12.0
flask
nltk
text-generation
opencv-python

argparse
colorama
dashscope
pyshine
pyyaml
requests
Levenshtein
zhipuai
xmltodict
google.auth